pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nova-network-failure-prediction"
version = "0.1.0"
description = "End-to-end pipeline for network failure prediction using Cisco telemetry + labeled flow captures + outage labeling heuristics."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Omar"}]
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyarrow>=14.0",
  "scikit-learn>=1.3",
  "torch>=2.0",
  "matplotlib>=3.7",
  "pyyaml>=6.0",
  "rich>=13.0",
  "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
